fix: 예외처리 수정
게임 종료시 (0 입력시) 오출력되는 부분 수정
잘못된 입력 처리후 게임 진행시 게임결과 부분 오출력되는 오류 수정
